% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calculate_LPI.R
\name{calculate_LPI}
\alias{calculate_LPI}
\title{Calculate the Living Planet Index for observations extracted from Atlas}
\usage{
calculate_LPI(target_taxa = "Tous")
}
\arguments{
\item{target_taxa}{}
}
\value{
lpi: a dataframe containing years, LPI values, and lower and upper 95% confidence intervals
}
\description{
Calculate the Living Planet Index for observations extracted from Atlas
}
