% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lpi_population.R
\name{lpi_population}
\alias{lpi_population}
\title{Calculate log-ratio growth rates and LPI trend at the population level}
\usage{
lpi_population(x)
}
\arguments{
\item{x}{Data frame of Atlas observations table joined with taxa table.}
}
\value{
Data frame 'x' with new columns \code{dt} which holds a list of log-ratio growth rates, \code{mean_dt} which holds the arithmetic mean log-ratio growth rate for the time series, and \code{lpi} which holds a list of index values calculated from the growth rates (dt).
}
\description{
Calculates the log-ratio change in raw population abundances for each population time series in the obs dataframe and calculates the (LPI) index values for the time series.
}
