% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_indextrend.R
\name{make_indextrend}
\alias{make_indextrend}
\title{Plot the Living Planet Index (LPI) trend}
\usage{
make_indextrend(target_taxa = "Tous")
}
\arguments{
\item{taxa}{can take the inputs "tous" (all populations in Québec), "amphibiens" , "mammifères", "oiseaux", "poissons", "reptiles" from the Shiny user's input.}
}
\value{
A plotly of the Living Planet Index trend through time, with the selected taxonomic group highlighted and shown with confidence intervals.
}
\description{
Function to plot each taxonomic group's LPI trend through time, where all unselected groups' trends are plotted in grey in the background, while the selected taxonomic group is plotted in colour and with its confidence intervals. There is also a plotly functionality to see the LPI value for each year when hovering over the plot.
}
