% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pointmap.R
\name{make_pointmap}
\alias{make_pointmap}
\title{Functions to make a map of population points that are in the dataset, which can be filtered by selected taxonomic group.}
\usage{
make_pointmap()
}
\arguments{
\item{target_taxa}{}
}
\value{
Leaflet map of point coordinations for populations included in the dataset.
}
\description{
Functions to make a map of population points that are in the dataset, which can be filtered by selected taxonomic group.
}
