% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_poptrend.R
\name{make_poptrend}
\alias{make_poptrend}
\title{Plot each population growth rate since 1990}
\usage{
make_poptrend(taxachoice)
}
\arguments{
\item{taxachoice}{}
}
\value{
A "bubbleplot" using \code{plotly} where each population is shown as a filled circle, called a "bubble" here. Each bubble's fill colour is scaled to the population's population growth rate (%). Bubbles are arranged on the \code{x} axis according to the population's growth rate (%) since 1990. The \code{y} axis is a set of random positions to differentiate the circles more easily. On hover, the common name of the population and a sentence explaining this population change rate (%) appears over the circle.
}
\description{
Makes a plot to compare individual populations' growth rates over the entire time frame of the estimated LPI trend (1990-2018).
}
