% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ridgeplot.R
\name{make_ridgeplot}
\alias{make_ridgeplot}
\title{Ridgeplot of the density of population growth rates per taxa}
\usage{
make_ridgeplot(taxachoice)
}
\value{
A ridgeplot of the density of all population growth rates for each taxa, stacked vertically to allow comparison between groups. The plot is themed with \code{theme_mapselector()}.
}
\description{
Ridgeplot of the density of population growth rates per taxa
}
