% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rlpi_population.R
\name{make_rlpi_population}
\alias{make_rlpi_population}
\title{Calculate LPI trend at the population level from rlpi outputs}
\usage{
make_rlpi_population(target_taxa)
}
\arguments{
\item{target_taxa}{}
}
\value{
Data frame containing years, LPI_final value, and population_id.
}
\description{
Used to plot population trends behind the mean LPI index trend, to communicate uncertainty.
}
