% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species_photo.R
\name{make_species_photo}
\alias{make_species_photo}
\title{Get species photo from iNaturalist}
\usage{
make_species_photo(clicked_population)
}
\arguments{
\item{clicked_population}{Population selected from a user's click on the point map.}
}
\value{
A photo of the selected population's species.
}
\description{
Gets a photo for a selected population's species name from iNaturalist. This function can then be used within the \code{summarise_rawdata} module to generate the species photo on the dashboard in reaction to a click.
}
