% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarise_rawdata.R
\name{make_summarise_rawdata}
\alias{make_summarise_rawdata}
\title{Summarise time series data into a descriptive table}
\usage{
make_summarise_rawdata(clicked_population)
}
\arguments{
\item{clicked_population}{Population selected from a user's click on the point map.}
}
\value{
A table describing the raw time series data.
}
\description{
Creates a table for a clicked population on the point map. This function can then be used within the \code{summarise_rawdata} module to generate the table on the dashboard in reaction to a click.
}
