% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_timeseries.R
\name{make_timeseries}
\alias{make_timeseries}
\title{Plot raw time series}
\usage{
make_timeseries(clicked_population)
}
\arguments{
\item{clicked_population}{Population selected from a user's click on the point map.}
}
\value{
A plotly of the raw time series of population size through time, with years on the \code{x} axis and population size on the \code{y} axis. Points are coloured following the same palette as the rest of the dashboard, based on taxa groups, and are linked by lines.
}
\description{
Plots a raw time series for a clicked population on the point map using \code{ggplot2}. This function can then be used within the \code{timeseries} module to generate the plot on the dashboard in reaction to a click.
}
