% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_trend_distribution.R
\name{make_trend_distribution}
\alias{make_trend_distribution}
\title{trend_distribution of the density of population growth rates per taxa}
\usage{
make_trend_distribution(target_year)
}
\value{
A plotly barchart showing the trend distribution of the density of all population growth rates for each taxa, stacked vertically to allow comparison between groups. The plot is themed with \code{theme_mapselector()}.
}
\description{
trend_distribution of the density of population growth rates per taxa
}
