% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pointmap.R
\docType{data}
\name{pal}
\alias{pal}
\title{Functions to make a map of population points that are in the dataset, which can be filtered by selected taxonomic group.}
\format{
An object of class \code{character} of length 6.
}
\usage{
pal
}
\arguments{
\item{target_taxa}{}
}
\value{
Leaflet map of point coordinations for populations included in the dataset.
}
\description{
Functions to make a map of population points that are in the dataset, which can be filtered by selected taxonomic group.
}
\keyword{datasets}
