% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarise_index_data.R
\name{summarise_index_data}
\alias{summarise_index_data}
\title{Summarise the dataset used to calculate the index.}
\usage{
summarise_index_data(target_taxa)
}
\arguments{
\item{target_taxa}{}
}
\value{
A text output of summary information about the quantity and characteristics of species included in the index.
}
\description{
Summarise the dataset used to calculate the index.
}
