% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-theme.R
\name{theme_mapselector}
\alias{theme_mapselector}
\title{Mapselector ggplot2 theme}
\usage{
theme_mapselector()
}
\value{
A \code{ggplot} theme that can be added to any \code{ggplot} object to change its theme. Plot titles are set to be larger, grid lines are adjusted to show only major grid lines, and facet colours and label size are set.
}
\description{
Mapselector ggplot2 theme
}
\examples{
library(ggplot2)
ggplot(diamonds) + 
geom_point(aes(x = carat, y = price, col = color)) + 
theme_mapselector()
}
